+++ sql/my_json_writer.cc
@@ -13,7 +13,7 @@
    along with this program; if not, write to the Free Software
    Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02111-1301 USA */
 
-#include <my_global.h>
+#include "mariadb.h"
 #include "sql_priv.h"
 #include "sql_string.h"
 
@@ -221,7 +221,7 @@ bool Single_line_formatting_helper::on_add_member(const char *name)
       buf_ptr+=len;
       *(buf_ptr++)= 0;
 
-      line_len= owner->indent_level + len + 1;
+      line_len= owner->indent_level + (uint)len + 1;
       state= ADD_MEMBER;
       return true; // handled
     }
@@ -286,7 +286,7 @@ bool Single_line_formatting_helper::on_add_str(const char *str)
     memcpy(buf_ptr, str, len);
     buf_ptr+=len;
     *(buf_ptr++)= 0;
-    line_len += len + 4;
+    line_len += (uint)len + 4;
     return true; // handled
   }
