+++ storage/myisam/mi_test1.c
@@ -16,6 +16,7 @@
 
 /* Testing of the basic functions of a MyISAM table */
 
+#include <my_global.h>
 #include "myisam.h"
 #include <my_getopt.h>
 #include <m_string.h>
@@ -385,7 +386,7 @@ static void create_key(uchar *key,uint rownr)
   }
   if (keyinfo[0].seg[0].flag & (HA_BLOB_PART | HA_VAR_LENGTH_PART))
   {
-    uint tmp;
+    size_t tmp;
     create_key_part(key+2,rownr);
     tmp=strlen((char*) key+2);
     int2store(key,tmp);
@@ -410,7 +411,7 @@ static void create_record(uchar *record,uint rownr)
   pos=record+1;
   if (recinfo[1].type == FIELD_BLOB)
   {
-    uint tmp;
+    size_t tmp;
     uchar *ptr;
     create_key_part(blob_key,rownr);
     tmp=strlen((char*) blob_key);
@@ -421,7 +422,7 @@ static void create_record(uchar *record,uint rownr)
   }
   else if (recinfo[1].type == FIELD_VARCHAR)
   {
-    uint tmp, pack_length= HA_VARCHAR_PACKLENGTH(recinfo[1].length-1);
+    size_t tmp, pack_length= HA_VARCHAR_PACKLENGTH(recinfo[1].length-1);
     create_key_part(pos+pack_length,rownr);
     tmp= strlen((char*) pos+pack_length);
     if (pack_length == 1)
@@ -437,7 +438,7 @@ static void create_record(uchar *record,uint rownr)
   }
   if (recinfo[2].type == FIELD_BLOB)
   {
-    uint tmp;
+    size_t tmp;
     uchar *ptr;;
     sprintf((char*) blob_record,"... row: %d", rownr);
     strappend((char*) blob_record,MY_MAX(MAX_REC_LENGTH-rownr,10),' ');
@@ -448,7 +449,7 @@ static void create_record(uchar *record,uint rownr)
   }
   else if (recinfo[2].type == FIELD_VARCHAR)
   {
-    uint tmp, pack_length= HA_VARCHAR_PACKLENGTH(recinfo[1].length-1);
+    size_t tmp, pack_length= HA_VARCHAR_PACKLENGTH(recinfo[1].length-1);
     sprintf((char*) pos+pack_length, "... row: %d", rownr);
     tmp= strlen((char*) pos+pack_length);
     if (pack_length == 1)
